+++ miosix/arch/drivers/sdmmc/stm32f2_f4_f7_sd.cpp
@@ -1656,7 +1656,7 @@ int SDIODriver::ioctl(int cmd, void* arg)
     }
     case IOCTL_GET_VOLUME_SIZE: {
         DBG("IOCTL_GET_VOLUME_SIZE\n");
-        off_t* sizePtr=static_cast<off_t*>(arg);
+        unsigned long long* sizePtr=static_cast<unsigned long long*>(arg);
         if (sizePtr==nullptr) {
             return -EINVAL;
         }
